requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Session==0.3.1
Flask-Caching==1.4.0
Flask-Migrate==2.3.0
Flask-Script==2.0.6
SQLAlchemy==1.2.13
SQLAlchemy-Utils==0.33.6
passlib==1.7.1
bcrypt==3.1.4
six==1.11.0
itsdangerous==1.1.0
requests>=2.20.0
PyMySQL==0.9.2
gunicorn==19.9.0
dataset==1.1.0
mistune==0.8.4
netaddr==0.7.19
redis==2.10.6
datafreeze==0.1.0
gevent==1.3.7
python-dotenv==0.9.1
flask-restplus==0.12.1
pathlib2==2.3.2
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.15.0
Flask-SocketIO==3.0.2
boto3==1.9.39
